+++ packages/jest-circus/src/__tests__/circus_it_test_error.test.js
@@ -33,21 +33,21 @@ describe('test/it error throwing', () => {
       circusIt('test1', () => {});
     }).not.toThrowError();
   });
-  it(`it throws error with missing callback function`, () => {
+  it(`it doesn't throw error with missing callback function`, () => {
     expect(() => {
-      // $FlowFixMe: Easy, we're testing runitme errors here
+      // $FlowFixMe: Easy, we're testing runtime errors here
       circusIt('test2');
-    }).toThrowError('Missing second argument. It must be a callback function.');
+    }).not.toThrowError();
   });
   it(`it throws an error when first argument isn't a string`, () => {
     expect(() => {
-      // $FlowFixMe: Easy, we're testing runitme errors here
+      // $FlowFixMe: Easy, we're testing runtime errors here
       circusIt(() => {});
     }).toThrowError(`Invalid first argument, () => {}. It must be a string.`);
   });
   it('it throws an error when callback function is not a function', () => {
     expect(() => {
-      // $FlowFixMe: Easy, we're testing runitme errors here
+      // $FlowFixMe: Easy, we're testing runtime errors here
       circusIt('test4', 'test4b');
     }).toThrowError(
       `Invalid second argument, test4b. It must be a callback function.`,
@@ -58,21 +58,21 @@ describe('test/it error throwing', () => {
       circusTest('test5', () => {});
     }).not.toThrowError();
   });
-  it(`test throws error with missing callback function`, () => {
+  it(`test doesn't throw error with missing callback function`, () => {
     expect(() => {
-      // $FlowFixMe: Easy, we're testing runitme errors here
+      // $FlowFixMe: Easy, we're testing runtime errors here
       circusTest('test6');
-    }).toThrowError('Missing second argument. It must be a callback function.');
+    }).not.toThrowError();
   });
   it(`test throws an error when first argument isn't a string`, () => {
     expect(() => {
-      // $FlowFixMe: Easy, we're testing runitme errors here
+      // $FlowFixMe: Easy, we're testing runtime errors here
       circusTest(() => {});
     }).toThrowError(`Invalid first argument, () => {}. It must be a string.`);
   });
   it('test throws an error when callback function is not a function', () => {
     expect(() => {
-      // $FlowFixMe: Easy, we're testing runitme errors here
+      // $FlowFixMe: Easy, we're testing runtime errors here
       circusTest('test8', 'test8b');
     }).toThrowError(
       `Invalid second argument, test8b. It must be a callback function.`,